Add a sphinx extension that does text substitution in docstrings.
The extension is used to replace "ASCII math" formulas in the boxcox and
boxcox1p docstring with math directives containing latex representations
of the formulas.

This allows the text shown, for example, in the docstring in ipython
to remain the readable ASCII text, while in the HTML output, the formula
is rendered nicely with mathjax.